+++ spec/kettle/dev/executable_version_spec.rb
@@ -22,4 +22,18 @@ RSpec.describe Kettle::Dev::ExecutableVersion do
       expect(described_class.header("kettle-example")).to eq("== kettle-example v#{Kettle::Dev::Version::VERSION} ==")
     end
   end
+
+  describe ".print_header" do
+    it "does not print by default" do
+      expect { described_class.print_header("kettle-example", []) }.not_to output.to_stdout
+    end
+
+    it "prints when verbose output is requested" do
+      argv = ["--verbose"]
+
+      expect { described_class.print_header("kettle-example", argv) }
+        .to output("== kettle-example v#{Kettle::Dev::Version::VERSION} ==\n").to_stdout
+      expect(argv).to eq([])
+    end
+  end
 end