+++ CHANGELOG.md
@@ -22,6 +22,9 @@ Please file a bug if you notice a violation of semantic versioning.
 
 ### Changed
 
+- Executable startup headers are now shown only when `--verbose` is passed;
+  `-v` and `--version` still print just the executable version and exit.
+
 ### Deprecated
 
 ### Removed